+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support Kyverno API v2beta1 schema.
+
 ## [1.26.1] - 2024-10-01
 
 ### Fixed